% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_events_rate.R
\name{format.power_events_rate}
\alias{format.power_events_rate}
\title{Format method for power_events_rate class}
\usage{
\method{format}{power_events_rate}(x, digits = 1, ...)
}
\arguments{
\item{x}{an R object of class power_events_rate}

\item{digits}{a positive integer indicating how many decimal digits are
to be used to display the probability columns as percentages.}

\item{...}{further arguments passed to or from other methods}
}
\value{
A character string with a markdown-style table of the
probabilities of observing each event threshold, by sample size and risk.
}
\description{
Format method for power_events_rate class
}
