% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_single_rate.R
\name{format.power_single_rate}
\alias{format.power_single_rate}
\title{Format method for power_single_rate class}
\usage{
\method{format}{power_single_rate}(x, digits = 3, ...)
}
\arguments{
\item{x}{an R object of class power_single_rate}

\item{digits}{a positive integer indicating how many significant digits are
to be used for numeric  x.}

\item{...}{further arguments passed to or from other methods}
}
\value{
A character string with a human-readable summary of the detection
power or a markdown-style table, depending on the number of rows.
}
\description{
Format method for power_single_rate class
}
