% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_power_best_binomial.R
\name{ggplot_lf_config}
\alias{ggplot_lf_config}
\title{Plot Least Favorable Configuration}
\usage{
ggplot_lf_config(x)
}
\arguments{
\item{x}{An object of class \code{lf_config}.}
}
\value{
A ggplot object.
}
\description{
Generates a ggplot showing empirical and predicted power to select the best
group over a grid of probabilities.
}
\examples{
\dontrun{
config <- lf_config(d = 0.2, k = 5, n = 25, simul = 1000)
ggplot_lf_config(config)
}
}
