% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_power_best_binomial.R
\name{ggplot_lf_config_binomial}
\alias{ggplot_lf_config_binomial}
\title{Plot Least Favorable Configuration}
\usage{
ggplot_lf_config_binomial(x)
}
\arguments{
\item{x}{An object of class \code{lf_config_binomial}.}
}
\value{
A ggplot object.
}
\description{
Generates a ggplot showing empirical and predicted power to select the best
group over a grid of probabilities.
}
\examples{
\dontrun{
config <- lf_config_binomial(dif = 0.2, ngroups = 5, npergroup = 25, nsim = 1000)
ggplot_lf_config_binomial (config)
}
}
