% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_power_best_binomial.R
\name{ggplot_prob_lowest_power}
\alias{ggplot_prob_lowest_power}
\title{GGplot of a prob_lowest_power object}
\usage{
ggplot_prob_lowest_power(x)
}
\arguments{
\item{x}{an object of prob_lower_power class}
}
\value{
a ggplot2 object
}
\description{
GGplot of a prob_lowest_power object
}
\examples{
\dontrun{
require(dplyr)
require(ggplot2)
lpb <- lowest_prop_best_binomial(dif = 0.2, ngroups = 5, npergroup = 25, nsimul = 10000)
ggplot_prob_lowest_power(lpb)
}
}
