% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_power_best_bin_rank.R
\name{ggplot_prob_lowest_power_bin_rank}
\alias{ggplot_prob_lowest_power_bin_rank}
\title{Plot Power Curve from a \code{prob_lowest_power_bin_rank} Object}
\usage{
ggplot_prob_lowest_power_bin_rank(x)
}
\arguments{
\item{x}{An object of class \code{prob_lowest_power_bin_rank}.}
}
\value{
A ggplot2 object.
}
\description{
Generates a ggplot showing simulated and fitted power curves for selecting
the best group based on ranks.
}
