% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_power_result.R
\name{is.empirical_power_result}
\alias{is.empirical_power_result}
\title{Check if an object is a sim_power_result}
\usage{
is.empirical_power_result(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
Logical. TRUE if \code{x} inherits from \code{"sim_power_result"}.
}
\description{
Check if an object is a sim_power_result
}
