% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_power_best_binomial.R
\name{lf_config}
\alias{lf_config}
\title{Least Favorable Configuration in a Binomial Experiment}
\usage{
lf_config(d, k, n, simul, prob = seq(0.05, 0.95, by = 0.01))
}
\arguments{
\item{d}{Numeric. Difference between best and other groups.}

\item{k}{Integer. Number of groups.}

\item{n}{Integer. Sample size per group.}

\item{simul}{Integer. Number of simulations.}

\item{prob}{Numeric vector. Grid of probabilities to evaluate.}
}
\value{
An S3 object of class \code{lf_config} with simulation results and fitted curve.
}
\description{
Identifies the event probability in the best group that results in the
lowest power to correctly detect it as best, using simulation and quadratic
model fitting.
}
\examples{
lf_config(d = 0.2, k = 4, n = 20, simul = 1000)
}
