% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_power_best_bin_rank.R
\name{lf_config_bin_rank}
\alias{lf_config_bin_rank}
\title{Least Favorable Configuration in a Binomial Rank Experiment}
\usage{
lf_config_bin_rank(
  noutcomes,
  dif,
  weights,
  ngroups,
  npergroup,
  nsim,
  p1 = seq(0.01, 0.99, length.out = 50)
)
}
\arguments{
\item{noutcomes}{Integer. Number of outcomes.}

\item{dif}{Numeric. Difference in probabilities between best and other groups.}

\item{weights}{Numeric vector. Weights for each outcome.}

\item{ngroups}{Integer. Number of groups.}

\item{npergroup}{Integer. Sample size per group.}

\item{nsim}{Integer. Number of simulations.}

\item{p1}{Optional vector of probabilities for the best group. If missing,
defaults to a sequence from 0.01 to 0.99.}
}
\value{
An S3 object of class \code{prob_lowest_power_bin_rank} with:
\itemize{
\item \code{minprob}: probability in best group with lowest power
\item \code{minpow}: estimated lowest power
\item \code{simulation}: simulation results with predicted curve
}
}
\description{
Identifies the event probability in the best group that results in the
lowest power to correctly ranked it as best in binomial experiment,
using simulation and quadratic model fitting.
}
