% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_power_best_binomial.R
\name{lf_config_binomial}
\alias{lf_config_binomial}
\title{Least Favorable Configuration in a Binomial Experiment}
\usage{
lf_config_binomial(
  dif,
  ngroups,
  npergroup,
  nsim,
  prob = seq(0.05, 0.95, by = 0.01)
)
}
\arguments{
\item{dif}{Numeric. Difference between best and other groups.}

\item{ngroups}{Integer. Number of groups.}

\item{npergroup}{Integer. Sample size per group.}

\item{nsim}{Integer. Number of simulations.}

\item{prob}{Numeric vector. Grid of probabilities to evaluate.}
}
\value{
An S3 object of class \code{lf_config_binomial} with simulation results and fitted curve.
}
\description{
Identifies the event probability in the best group that results in the
lowest power to correctly detect it as best, using simulation and quadratic
model fitting.
}
\examples{
lf_config_binomial(dif = 0.2, ngroups = 4, npergroup = 20, nsim = 100)
}
