% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multt.R, R/multt_b.R
\name{multt}
\alias{multt}
\title{Calculate Upper Equicoordinate Point for a Multivariate t Distribution}
\usage{
multt(alpha, n, rho, df, seed = NULL)

multt(alpha, n, rho, df, seed = NULL)
}
\arguments{
\item{alpha}{the significance level (e.g., 0.05 for a 95\% confidence level).}

\item{n}{the number of variables in the multivariate normal distribution.}

\item{rho}{the common correlation coefficient between the variables.}

\item{df}{degrees of freedom for the t distribution}

\item{seed}{an object specifying if and how the random number generator should be initialized. Used by the function \code{qmvt}}
}
\value{
The upper equicoordinate point \eqn{z} such that the probability of all
variables being less than or equal to \eqn{z} is \eqn{1 - \alpha}.

The upper equicoordinate point \eqn{z} such that the probability of all
variables being less than or equal to \eqn{z} is \eqn{1 - \alpha}.
}
\description{
This function computes the upper equicoordinate point for a multivariate t
distribution with unit variances and a common correlation coefficient.
The result is the quantile such that the joint cumulative probability equals \eqn{1 - \alpha}, single tail.

This function computes the upper equicoordinate point for a multivariate t
distribution with unit variances and a common correlation coefficient.
The result is the quantile such that the joint cumulative probability equals \eqn{1 - \alpha}, double tails
}
\examples{
alpha <- 0.1  # Significance level (10\%)
n <- 3        # Number of variables
rho <- 0.5    # Common correlation coefficient
df = 5
multt(alpha, n, rho, df)
alpha <- 0.1  # Significance level (10\%)
n <- 3        # Number of variables
rho <- 0.5    # Common correlation coefficient
df = 5
multt(alpha, n, rho, df)
}
