% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_best_bin_binomial.R
\name{n_best_bin_binomial}
\alias{n_best_bin_binomial}
\title{Sample size for selection of the best group in a binomial test}
\usage{
n_best_bin_binomial(power, p1, dif, ngroups, max_n = 1000)
}
\arguments{
\item{power}{target power}

\item{p1}{probability of the best group}

\item{dif}{difference in probability with the next best group}

\item{ngroups}{number of groups}

\item{max_n}{maximum n evaluated}
}
\description{
Find the minimum sample size per group with a \code{power}
to correctly classify the best group as best from \code{ngroups} groups if the true
probability of events in best group is \code{p1} and the difference
with the next group is \code{difs}
}
