% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_best_bin_binomial.R
\name{power_best_bin_binomial}
\alias{power_best_bin_binomial}
\title{Power to correctly classify as best the best group in a binomial test}
\usage{
power_best_bin_binomial(p1, dif, ngroups, npergroup)
}
\arguments{
\item{p1}{probability in the best group}

\item{dif}{Difference between the best group and the next}

\item{ngroups}{Number of groups}

\item{npergroup}{Number of subjects in each group}
}
\value{
the probability of correctly classify as best the best group
}
\description{
Estimate the exact probability of correctly classify as best the best group when
the outcome follows a binomial distribution, assuming that \code{p1} is the
probability of events in the best group, and the probability in the next
group is \code{dif} lower
}
\examples{
power_best_bin_binomial( p1 = 0.8, dif = 0.2, ngroups = 4, npergroup = 50)
}
\references{
Sobel, Milton, and Marilyn J. Huyett. “Selecting the Best One of
Several Binomial Populations.” Bell System Technical Journal 36, no. 2
(1957): 537–76. https://doi.org/10.1002/j.1538-7305.1957.tb02411.x.
}
