% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_best_bin_rank.R
\name{power_best_bin_rank}
\alias{power_best_bin_rank}
\title{Power to select the best group}
\usage{
power_best_bin_rank(noutcomes, prob, dif, ngroups, npergroup, nsimul)
}
\arguments{
\item{noutcomes}{number of outcomes to evaluate}

\item{prob}{the probability  in the rest of the groups for each outcome}

\item{dif}{difference between the most promising and the rest of the groups for each outcome}

\item{ngroups}{number of groups to compare}

\item{npergroup}{number of subjects in each group}

\item{nsimul}{number of simulations}
}
\value{
The empirical power expressed as the proportion of the simulations
where the best group was selected as best for all outcomes.
}
\description{
This function estimates the empirical power to select the most promising group as
the best group, when the outcome follows a binomial distribution, assuming that
the probability of the most promising group is at least \code{dif} higher than the least promising groups.
}
\details{
The function allows to evaluate several outcomes at the same time, in which
case evaluates if the most promising group is the best for all outcomes.

It assume each outcome follows a binomial distribution and they are independent between them.
The outcomes may have the same probability and same difference between the most promising
group and the others, or they can be defined differently by outcome. Differently
from power_best_binomial, in which the
the most promising has to be the best for all outcomes,
this simulation consider the sum of the ranks for each test and return
the power to the first group to be selected as the best

The number of subjects per group can be the same or can be specified
for each group. If specified, the most promising group is always the first group.
}
\examples{
\dontrun{
# Power to select the best group if the difference between the best and
# the the other two groups is 0.2. One outcome and three groups of 30 subjects
#
power_best_bin_rank(
  noutcomes = 1,
  prob = 0.5,
  dif = 0.2,
  ngroups= 3,
  npergroup= 30,
  nsimul=10000

}
}
