% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_equivalence_normal.R
\name{power_equivalence_normal}
\alias{power_equivalence_normal}
\title{Empirical power for equivalence}
\usage{
power_equivalence_normal(
  ngroups,
  npergroup,
  sd,
  llimit,
  ulimit,
  nsimul,
  conf.level = 0.95
)
}
\arguments{
\item{ngroups}{number of groups to compare}

\item{npergroup}{number of observations per group}

\item{sd}{standard deviation of the distribution}

\item{llimit}{lower limit for equivalence}

\item{ulimit}{upper limit for equivalence}

\item{nsimul}{number of simulations}

\item{conf.level}{confidence level of the interval}
}
\value{
the power to declare equivalence
}
\description{
Estimate the empirical power to detect equivalence assuming a true
no difference in normally distributed outcomes, using pairwise t.test
comparisons between the different groups. Equivalence is declared if the
confidence interval of the difference are within the lower limit (\code{llimit})
and the upper limit (\code{ulimit})
}
\examples{
# Lot to lot equivalence for three lots, assuming a standard deviation
# in log10 scale of 0.402 and limits for equivalence log10(2/3) and log10(3/2)

power_equivalence_normal(
ngroups = 3,
npergroup = 172,
sd = 0.403,
llimit = log10(2/3),
ulimit = log10(3/2),
nsimul = 1000,
conf.level = 0.95
)
}
