% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_non_inferiority.R
\name{power_ni_normal}
\alias{power_ni_normal}
\title{Empirical power for non-inferiority}
\usage{
power_ni_normal(
  ntest,
  npergroup,
  sd,
  ni_limit,
  test_req,
  test_opt,
  nsimul,
  conf.level = 0.95
)
}
\arguments{
\item{ntest}{number of test to compare}

\item{npergroup}{number of observations per group}

\item{sd}{standard deviation of the distribution. can be a single number for all test or a vector with sd for each test}

\item{ni_limit}{limit to declare non-inferiority of a test. It can be a single number of a vector with limits for each test}

\item{test_req}{Number of test required to accomplish non-inferiority. Always the first tests}

\item{test_opt}{Number of the test that must accomplish non-inferiority from the rest of
the test non required}

\item{nsimul}{number of simulations}

\item{conf.level}{confidence level of the interval. One number for all test or a vector to specify a different conf.level for each test}
}
\value{
a tibble with the power to declare non inferiority and 95\% CI
}
\description{
Estimate the empirical power to detect non-inferiority assuming a true
no difference in normally distributed outcomes, using t.test for several tests.
Non-inferiority for a single test is defined as having a lower confidence
interval of the comparison is higher than the non-inferiority limit.
Overall non-inferiority is declared as a rule for specific test or a number
of test from the group (ie 3 of the five are non-inferior)
}
\examples{
# Non-Inferiority for 7 test, from which the first two must accomplish
# non-inferiority and 3 of the left 5 must accomplish non inferiority
# sd and ni_limit is the same for the 7 tests. sd is in log10 scale

power_ni_normal(
ntest = 7,
npergroup = 250,
sd = 0.4,
ni_limit = log10(2/3),
test_req = 2,
test_opt = 3,
nsimul = 1000,
conf.level = 0.95
)
}
