% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_single_rate.R
\name{power_single_rate}
\alias{power_single_rate}
\title{Rate of events to observe at least one event}
\usage{
power_single_rate(subjects, power)
}
\arguments{
\item{subjects}{sample size}

\item{power}{power}
}
\value{
and s3 object of class power_single_rate with the proportion that can be
detected with the specified power and sample size
}
\description{
Estimate which is the true proportion required to observe
at least one event given a sample size and a power
}
\examples{
power_single_rate(30,0.9)
power_single_rate(c(30,50,100), c(0.9))
}
