% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_power_result.R
\name{print.empirical_power_result}
\alias{print.empirical_power_result}
\title{Print method for empirical_power_result}
\usage{
\method{print}{empirical_power_result}(x, ...)
}
\arguments{
\item{x}{An object of class "empirical_power_result".}

\item{...}{Further arguments passed to or from other methods (ignored).}
}
\description{
Nicely formats the output of an object of class \code{empirical_power_result},
showing the power estimate, confidence interval, and number of simulations.
}
