% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_events_rate.R
\name{print.power_events_rate}
\alias{print.power_events_rate}
\title{Print method for class power_events_rate}
\usage{
\method{print}{power_events_rate}(x, ...)
}
\arguments{
\item{x}{an object of class power_events_rate}

\item{...}{further arguments passed to or from other methods}
}
\value{
Invisibly returns the object passed in.
}
\description{
Print method for class power_events_rate
}
