% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_power.R
\name{rate_power}
\alias{rate_power}
\title{Rate for a given power}
\usage{
rate_power(subjects, power)
}
\arguments{
\item{subjects}{sample size}

\item{power}{power}
}
\value{
and s3 object of class rate_power with the proportion that can be
detected with the specified power and sample size
}
\description{
Estimate which is the true proportion required to observe
at least one event given a sample size and a power
}
\examples{
rate_power(30,0.9)
rate_power(c(30,50,100), c(0.9))
}
