% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usenb.R
\name{ss_usenb}
\alias{ss_usenb}
\title{Sample Size for Selecting the Best Treatment in a Single-Factor Normal Response Using the Indifference-Zone Approach}
\usage{
ss_usenb(groups, delta, sd, power, seed = NULL)
}
\arguments{
\item{groups}{number of groups in the test.}

\item{delta}{change to be detected.}

\item{sd}{common standard deviation of the response variable.}

\item{power}{desired power of the test.}

\item{seed}{the seed used for the multz procedure}
}
\value{
Returns the sample size needed for each group to achieve the desired power.
}
\description{
Estimate the minimum common sample size per group required to guarantee the
probability (power) of correctly selecting the best group
(i.e., the group with the highest mean) using the indifference-zone approach.
}
\details{
The procedure assumes a normal distribution of responses, with a common known
variance. 'delta' is the smallest difference considered worth detecting between
the best group and the next least favorable group.

Reference: Bechhofer, R.E., Santner, T.J., & Goldsman, D.M. (1995).
Design and Analysis of Experiments for Statistical Selection,
Screening, and Multiple Comparisons.
Wiley Series in Probability and Statistics.
}
\examples{
ss_usenb(groups = 3, delta = 0.5, sd = 1, power = 0.8)

}
