% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssutil-package.R
\docType{package}
\name{ssutil-package}
\alias{ssutil}
\alias{ssutil-package}
\title{ssutil: Utilities for Sample Size calculation}
\description{
Utilities for the sample size estimation and
simulations of clinical trials.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://johnaponte.github.io/ssutil/}
  \item \url{https://github.com/johnaponte/ssutil}
  \item Report bugs at \url{https://github.com/johnaponte/ssutil/issues}
}

}
\author{
\strong{Maintainer}: John J. Aponte \email{john.j.aponte@gmail.com} (\href{https://orcid.org/0000-0002-3014-3673}{ORCID})

Authors:
\itemize{
  \item John J. Aponte \email{john.j.aponte@gmail.com} (\href{https://orcid.org/0000-0002-3014-3673}{ORCID})
}

Other contributors:
\itemize{
  \item Chris Gast \email{cgast@path.org} [contributor]
}

}
\keyword{internal}
