% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_power_result.R
\name{tidy.empirical_power_result}
\alias{tidy.empirical_power_result}
\title{Tidy Method for empirical_power_result}
\usage{
\method{tidy}{empirical_power_result}(x, ...)
}
\arguments{
\item{x}{A \code{empirical_power_result} object.}

\item{...}{Ignored.}
}
\value{
A tibble with columns: \code{power}, \code{conf.low},
\code{conf.high},  \code{conf.level}. \code{nsim}.
}
\description{
Creates a one-row tibble with the power estimate and confidence interval.
}
