% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_power_best_binomial.R
\name{ggplot_lf_config}
\alias{ggplot_lf_config}
\title{GGplot least configuration simulation}
\usage{
ggplot_lf_config(x)
}
\arguments{
\item{x}{an object of lf_config class}
}
\value{
a ggplot2 object
}
\description{
GGplot least configuration simulation
}
\examples{
\dontrun{
lpb <- lf_config(d = 0.2, k = 5, n = 25, simul = 1000)
ggplot_lf_config(lpb)
}
}
