% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_power_best_binomial.R
\name{lf_config}
\alias{lf_config}
\title{Least favorable configuration for a binomial experiment}
\usage{
lf_config(d, k, n, simul, p1 = seq(from = 0.05, to = 0.95, by = 0.01))
}
\arguments{
\item{d}{indifference zone}

\item{k}{number of groups}

\item{n}{number of subjects per group}

\item{simul}{number of simulations for each scenario}

\item{prob}{grid of probabilities for the best group. By default is a sequence between 0.05 and 0.95 by 0.01}
}
\value{
an S3 object of class lf_config
}
\description{
Given a difference, a number of groups and a number of subjects per groups, use
Monte Carlo simulations to find the least favorable configuration to correctly
select the best group.
}
\details{
The program estimate a quadratic fit to the empirical power over a grid of
potential probabilities for the best group, and identify the probability with
the lowest power to correctly identify as best the best group.
}
\examples{
lf_config(d=0.2, k=4, n-20, simul=1000)

}
