% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_best_bin_rank.R
\name{lowest_prop_best_bin_rank}
\alias{lowest_prop_best_bin_rank}
\title{Probability with lowest power to detect a difference using ranks}
\usage{
lowest_prop_best_bin_rank(
  dif,
  ngroups,
  npergroup,
  nsimul,
  prob = seq(from = 0.05, to = 0.95, by = 0.01)
)
}
\arguments{
\item{dif}{difference between the best and the least promising groups}

\item{ngroups}{number of groups}

\item{npergroup}{number of subjects per group}

\item{nsimul}{number of simulations}

\item{prob}{probabilities on the least promising group to evaluate. By default is a sequence between 0.05 and 0.95 by 0.01}
}
\value{
an S3 object of class prob_lowest_power
}
\description{
Estimates the probability of the less promising groups with the
lowest power to detect a difference between groups using ranks
A simulation is made for a grid of probabilities and a quadratic
function is fitted to the estimated powers by probability. The
function returns the minimum probability from the quadratic function.
function.
It is restricted to 1 outcome.
}
