% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophr.R
\name{prophr}
\alias{prophr}
\title{Calculate the proportion of events given a hazard ratio}
\usage{
prophr(p0, hr)
}
\arguments{
\item{p0}{Probability of an event in the control group}

\item{hr}{Hazard ratio}
}
\value{
The probability of event in the experimental group
}
\description{
This calculate the proportion of events in the experimental group
based on the proportion of event in the control group and the hazard ratio,
assuming proportionality of the hazard assumption.
}
\details{
This is useful when evaluating sample size in PASS(TM) as it does not make the
correction of the other group automatically.
}
\examples{
prophr(0.05,0.6)
}
