% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_power_equivalence_normal.R
\name{sim_power_equivalence_normal}
\alias{sim_power_equivalence_normal}
\title{Empirical power for equivalence}
\usage{
sim_power_equivalence_normal(
  ngroups,
  npergroup,
  sd,
  llimit,
  ulimit,
  nsimul,
  conf.level = 0.95
)
}
\arguments{
\item{ngroups}{number of groups to compare}

\item{npergroup}{number of observations per group}

\item{sd}{standard deviation of the distribution}

\item{llimit}{lower limit for equivalence}

\item{ulimit}{upper limit for equivalence}

\item{nsimul}{number of simulations}

\item{conf.level}{confidence level of the interval}
}
\value{
a tibble with the power to declare equivalence and 95\% CI
}
\description{
Estimate the empirical power to detect equivalence assuming a true
no difference in normally distributed outcomes, using t.test on all pair
comparisons between the different groups. Equivalence is declared if the
confidence interval of the difference are within the lower limit (\code{llimit})
and the upper limit (\code{ulimit}) for all comparisons
}
\examples{
# Lot to lot equivalence for three lots, assuming a standard deviation
# in log10 scale of 0.402 and limits for equivalence log10(2/3) and log10(3/2)

sim_power_equivalence_normal(
ngroups = 3,
npergroup = 172,
sd = 0.403,
llimit = log10(2/3),
ulimit = log10(3/2),
nsimul = 1000,
conf.level = 0.95
)
}
