% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_power_equivalence_normal.R
\name{sim_power_equivalence_normal}
\alias{sim_power_equivalence_normal}
\title{Empirical Power for Equivalence (Normal Outcomes)}
\usage{
sim_power_equivalence_normal(
  ngroups,
  npergroup,
  sd,
  llimit,
  ulimit,
  nsimul,
  conf.level = 0.95
)
}
\arguments{
\item{ngroups}{Integer. Number of groups to compare}

\item{npergroup}{Integer. Number of observations per group.}

\item{sd}{Numeric. Standard deviation of the outcome distribution (common across groups).}

\item{llimit}{Numeric. Lower equivalence limit.}

\item{ulimit}{Numeric. Upper equivalence limit.}

\item{nsimul}{Integer. Number of simulations to perform.}

\item{conf.level}{Numeric. Confidence level used for the t-tests (e.g., 0.95 for 95\% CI).}
}
\value{
A data frame with the following columns:\tabular{ll}{
   Column \tab Description \cr
   power \tab Empirical power estimate. \cr
   conf.low \tab Lower bound of 95\% confidence interval \cr
   conf.high \tab Upper bound of 95\% confidence interval \cr
   nsim \tab Number of simulations performed \cr
}
}
\description{
Estimates the empirical power to detect equivalence among multiple groups assuming
no true difference in normally distributed outcomes. Pairwise two-sample t-tests are
used, and equivalence is declared if all confidence intervals for differences between
group means lie entirely within the interval defined by \code{llimit} and \code{ulimit}.
}
\details{
This function simulates data under the null hypothesis of no difference between groups
and calculates the proportion of simulations in which all pairwise comparisons fall within
the specified equivalence limits.
}
\examples{
\dontrun{
#Equivalence testing for three groups with log-scale outcome
sim_power_equivalence_normal(
  ngroups = 3,
  npergroup = 172,
  sd = 0.403,
  llimit = log10(2/3),
  ulimit = log10(3/2),
  nsimul = 1000,
  conf.level = 0.95
)
}
}
