% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_power_nbinom.R
\name{sim_power_nbinom}
\alias{sim_power_nbinom}
\title{Empirical power for negative binomial}
\usage{
sim_power_nbinom(n1, n2, ir1, tm, rr, lowrr, dispersion, alpha, nsimul)
}
\arguments{
\item{n1}{Number of participants in group 1}

\item{n2}{Number of participants in group 2}

\item{ir1}{Incidence rate in group 1}

\item{tm}{Average exposure time}

\item{rr}{Relative risk between group 2 and group 1 (ir2/ir1)}

\item{lowrr}{Lower limit of the Relative ratio target of the comparison}

\item{dispersion}{Dispersion parameter}

\item{alpha}{= Type 1 error (two tails)}

\item{nsimul}{Number of simulations}
}
\description{
Estimates the empirical power for negative binomial comparisons.
}
\note{
this uses the alternative nomenclature for negative binomial where
the size = 1/phi and prob = size/(size + mu) where mu is the mean parameter
which for group 1 is mu1 = ir1\*tm and for group 2 is mu2 = ir1\*rr\*tm.
size is estimated as theta in \code{glm.nb}
}
\author{
{Chris G}
}
