% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_ni_ve.R
\name{ss_ni_ve}
\alias{ss_ni_ve}
\title{Non inferiority for vaccine efficacy}
\usage{
ss_ni_ve(ve_lci, alpha = 0.025, power = 0.9, use70 = FALSE, preserve = 0.5)
}
\arguments{
\item{ve_lci}{Lower limit of the vaccine efficacy}

\item{alpha}{Error type 1}

\item{power}{Power}

\item{use70}{Assume new vaccine is at least 30\% instead of half of the current vaccine, otherwise half of the VE}

\item{preserve}{Preserve effect- By default 0.5}
}
\value{
a list with the non-inferiority margin, number of events and maximum
hazard ratio and events in the experimental to declare Non inferiority
}
\description{
This evaluate the non-inferiority limit, the number of events and
the maximum HR to declare inferiority following Fleming et al
COVID-19 vaccine trials:
The use of active controls and non-inferiority studies. Clinical Trials
18(3), 335-342. https://doi.org/10.1177/1740774520988244
following the 95-95 or 90-70 rule
}
\details{
Please not that is intended to produce Table 1 of the paper but the results
are not exactly similar, I think because rounding and the way the
sample 1 binomial is calculated, how the confidence intervals for the
proportion are estimated but it is closed enough. Here we use sample1binomial
from gsDesign and exact confidence intervals
}
